fix: socket erros during relocation
