+++ pool/options.go
@@ -15,9 +15,3 @@ func WithCustomErrEncoder(errEnc ErrorEncoder) Options {
 		p.errEncoder = errEnc
 	}
 }
-
-func UseParallelAlloc() Options {
-	return func(p *StaticPool) {
-		p.parallelAlloc = true
-	}
-}
